news_service/requirements.txt
Django==4.2.2
django-redis==5.3.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
psycopg2-binary==2.9.6
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3